refactor alibaba provider modules
